Removed the warnings with the currently active flags Squashed commit of the following:
commit 22674c905ca53770942c3c699d19c873b688c6aa
Author: Daniele Rapetti <5535617+Iximiel@users.noreply.github.com>
Date:   Fri Dec 1 10:49:35 2023 +0100

    -Werror on PlumedPythonEmbeddedModule

commit 0d6e623ecb238cd75aa8c84190075424c854cc42
Author: Daniele Rapetti <5535617+Iximiel@users.noreply.github.com>
Date:   Fri Dec 1 10:48:58 2023 +0100

    -Werror on PythonFunction

commit ff503e4791e6a64c04147f65b0bfc38155ae7b08
Author: Daniele Rapetti <5535617+Iximiel@users.noreply.github.com>
Date:   Fri Dec 1 10:34:54 2023 +0100

    -Werror on PythonCVInterface

+++ plugins/pycv/PythonCVInterface.cpp
@@ -671,7 +671,8 @@ void PythonCVInterface::readReturn(const py::object &r, Value* valPtr) {
     py::list rl=r.cast<py::list>();
     pycvComm_t value = rl[0].cast<pycvComm_t>();
     valPtr->set(value);
-    auto natoms = getPositions().size();
+    //shape returns long int
+    auto natoms = static_cast<long int > (getPositions().size());
     if (rl.size() > 1) {
       if(!valPtr->hasDerivatives())
         error(valPtr->getName()+" was declared without derivatives, but python returned with derivatives");
